Refactor code structure
This commit refactors out the registration into different classes, as
well as change the way dinosaurs are created, to make the creation of
them easier in the future. WARNING: To be consistent with vanilla, as
well as making my life easier in the future, this commit also changes
the registry name of all entities/items. This means that this update
breaks compatability with old worlds.